combine all tests to api_test.py
